# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Portfolio API",
        "description": "API documentation for the Portfolio and Blog project",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/api/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get list of categories",
                "description": "Returns list of categories",
                "operationId": "getCategoriesList",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/contact": {
            "post": {
                "tags": [
                    "Contact"
                ],
                "summary": "Store new contact message",
                "description": "Returns contact data",
                "operationId": "storeContact",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ContactRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api/posts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Get list of blog posts",
                "description": "Returns list of published blog posts",
                "operationId": "getPostsList",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/projects": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Get list of projects",
                "description": "Returns list of projects",
                "operationId": "getProjectsList",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/projects/{id}": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Get project information",
                "description": "Returns project data",
                "operationId": "getProjectById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Project id or slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/skills": {
            "get": {
                "tags": [
                    "Skills"
                ],
                "summary": "Get list of skills",
                "description": "Returns list of skills",
                "operationId": "getSkillsList",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "ContactRequest": {
                "title": "Contact Request",
                "description": "Contact form request body",
                "required": [
                    "name",
                    "email",
                    "message"
                ],
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "Name of the sender",
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "title": "email",
                        "description": "Email of the sender",
                        "type": "string",
                        "format": "email",
                        "example": "john@example.com"
                    },
                    "subject": {
                        "title": "subject",
                        "description": "Subject of the message",
                        "type": "string",
                        "example": "Project Inquiry"
                    },
                    "message": {
                        "title": "message",
                        "description": "Message content",
                        "type": "string",
                        "example": "I would like to hire you."
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Categories",
            "description": "Categories"
        },
        {
            "name": "Contact",
            "description": "Contact"
        },
        {
            "name": "Posts",
            "description": "Posts"
        },
        {
            "name": "Projects",
            "description": "Projects"
        },
        {
            "name": "Skills",
            "description": "Skills"
        }
    ]
}